fix todo: remove SharedHandler
